refactor(admin): relocate time-bucket helpers to utils; add growth rate layer
